client - user manager redesign
