+++ client/src/components/admin/users/comps/AdminUserComps.js
@@ -2,10 +2,10 @@ import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { isEmpty } from 'lodash';
 
-import { List, Header, Accordion } from 'semantic-ui-react';
+import { Header, Accordion } from 'semantic-ui-react';
 
 import AdminRoleReqComps from './AdminRoleReqComps';
-import AdminCurrentComps from './AdminCurrentComps';
+// import AdminCurrentComps from './AdminCurrentComps';
 
 import {
   selectUserManage,
@@ -33,12 +33,6 @@ const AdminUserComps = () => {
           <Accordion fluid styled>
             <AdminRoleReqComps user={user} />
           </Accordion>
-          <Header as="h4" textAlign="center">
-            Current Competencies
-          </Header>
-          <List divided verticalAlign="middle">
-            <AdminCurrentComps user={user} currentComps={currentComps} />
-          </List>
         </>
       )}
     </>
