+++ client/src/components/admin/users/courses/AdminUserCoursesCurrent.js
@@ -69,7 +69,7 @@ class AdminUserCoursesCurrent extends Component {
       return (
         <Table.Row key={_id}>
           <Table.Cell>{_course.courseName}</Table.Cell>
-          <Table.Cell textAlign="right">{validText}</Table.Cell>
+          <Table.Cell>{validText}</Table.Cell>
           <Table.Cell>{_course.type}</Table.Cell>
           <Table.Cell>{_course.level}</Table.Cell>
           <Table.Cell>
@@ -93,7 +93,7 @@ class AdminUserCoursesCurrent extends Component {
       <>
         {courses.length > 0 && (
           <>
-            <Header as="h4" textAlign="center">
+            <Header as="h3" textAlign="center">
               Current Courses
             </Header>
             <Table sortable celled fixed structured>
